Added Makefile
Useful for testing and cleaning up .pyc files and __pycache__
