+++ makefile
@@ -1,0 +1,10 @@
+SUBDIR_ROOTS := modules tests
+DIRS := . $(shell find $(SUBDIR_ROOTS) -type d)
+GARBAGE_PATTERNS := __pycache__
+GARBAGE := $(foreach DIR,$(DIRS),$(addprefix $(DIR)/,$(GARBAGE_PATTERNS)))
+
+clean:
+	rm -rf $(GARBAGE)
+
+test:
+	python3 -m pytest
