+++ tests/peripheral_tests/test_heap.py
@@ -2,28 +2,33 @@ import modules.peripherals.heap as heap
 import random
 
 #tests heap pop
-a  = []
-for i in range (10):
-    a.append(random.randint(0,100))
+def test_pop():
+    a  = []
+    for i in range (50):
+        a.append(random.randint(0,100))
 
-b = heap.heap()
-b.heapify(a)
+    b = heap.heap()
+    b.heapify(a)
 
-c = []
-for i in range(10):
-    element = b.pop()
-    c.append(element)
+    c = []
+    for i in range(50):
+        element = b.pop()
+        c.append(element)
 
-a.sort(reverse = True)
-assert a == c
+    a.sort(reverse = True)
+    assert a == c
 
 #tests heap push
-d = []
-for i in range(10):
-    d.append(random.randint(0,100))
+def test_push():
+    d = []
+    for i in range(50):
+        d.append(random.randint(0,100))
 
-for element in d:
-    b.push(element)
-    assert max(b.data) == b.data[1]
+    b = heap.heap()
+    b.heapify(d)
+
+    for element in d:
+        b.push(element)
+        assert max(b.data) == b.data[1]
